# server/openapi.yaml
openapi: 3.0.3
info:
  title: Campus Study Buddy API
  description: API for Campus Study Buddy - A platform for students to find study partners and form study groups
  version: 1.0.0
  contact:
    name: Campus Study Buddy Team
    email: support@campusstudybuddy.com
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.campusstudybuddy.com
    description: Production server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase ID token
  schemas:
...
    CalendarEvent:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        startISO:
          type: string
          format: date-time
        endISO:
          type: string
          format: date-time
        location:
          type: string
        attendees:
          type: array
          items:
            type: object
            properties:
              email:
                type: string
                format: email
              responseStatus:
                type: string
                enum: [needsAction, accepted, declined, tentative]
        creator:
          type: object
          properties:
            email:
              type: string
              format: email
            displayName:
              type: string
...
paths:
...
  /api/schedule:
    get:
      summary: Get calendar events
      description: Retrieve calendar events (mock implementation)
      tags:
        - Schedule
      security:
        - BearerAuth: []
      parameters:
        - name: timeMin
          in: query
          description: Start time filter (ISO 8601)
          required: false
          schema:
            type: string
            format: date-time
        - name: timeMax
          in: query
          description: End time filter (ISO 8601)
          required: false
          schema:
            type: string
            format: date-time
        - name: maxResults
          in: query
          description: Maximum number of events to return
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: List of calendar events
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/CalendarEvent'
                  count:
                    type: integer
                  total:
                    type: integer
                  note:
                    type: string
                  mock:
                    type: boolean
        '401':
          description: Unauthorized - Invalid or missing token
        '500':
          description: Internal server error

    post:
      summary: Create a calendar event
      description: Create a new calendar event (mock implementation)
      tags:
        - Schedule
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - startISO
                - endISO
              properties:
                title:
                  type: string
                  description: Event title
                  minLength: 1
                  maxLength: 200
                description:
                  type: string
                  description: Event description
                  maxLength: 500
                startISO:
                  type: string
                  format: date-time
                  description: Start time (ISO 8601)
                endISO:
                  type: string
                  format: date-time
                  description: End time (ISO 8601)
                location:
                  type: string
                  description: Event location
                attendees:
                  type: array
                  items:
                    type: string
                    format: email
                  description: List of attendee emails
      responses:
        '201':
          description: Event created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/CalendarEvent'
                  message:
                    type: string
                  note:
                    type: string
                  mock:
                    type: boolean
        '400':
          description: Bad request - Validation error
        '401':
          description: Unauthorized - Invalid or missing token
        '500':
          description: Internal server error
  /api/schedule/{id}:
    get:
      summary: Get a specific calendar event
      description: Retrieve details of a specific calendar event (mock implementation)
      tags:
        - Schedule
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Event ID
          schema:
            type: string
      responses:
        '200':
          description: Event details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/CalendarEvent'
                  note:
                    type: string
                  mock:
                    type: boolean
        '401':
          description: Unauthorized - Invalid or missing token
        '404':
          description: Event not found
        '500':
          description: Internal server error
...